requirements.txt
azure-functions
openai
azure-search-documents==11.4.0
python-dotenv